+++ src/Synapse/Synapse/Models/PSManagedIdentity.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.Azure.Management.Synapse.Models;
+
+namespace Microsoft.Azure.Commands.Synapse.Models
+{
+    public class PSManagedIdentity
+    {
+        public string IdentityType { get; set; }
+
+        public string PrincipalId { get; set; }
+
+        public string TenantId { get; set; }
+
+        public PSManagedIdentity(ManagedIdentity identity)
+        {
+            this.IdentityType = identity?.Type?.ToString();
+            this.PrincipalId = identity?.PrincipalId;
+            this.TenantId = identity?.TenantId;
+        }
+    }
+}