+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Azure Site Recovery added support for protecting proximity placement group virtual machines for Azure to Azure provider.
 
 ## Version 2.8.0
 * Azure Site Recovery added support for doing reprotect and updated vm properties for Azure disk encrypted Virtual Machines.